fixed check args problem of int

+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rdehouck <rdehouck@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: robindehouck <robindehouck@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 11:59:46 by rdehouck          #+#    #+#             */
-/*   Updated: 2022/02/08 11:46:33 by rdehouck         ###   ########lyon.fr   */
+/*   Updated: 2022/03/24 13:57:49 by robindehouc      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ int			ft_strncmp(const char *source1, const char *source2, size_t n);
 void		*ft_memchr(const void *source, int c, size_t n);
 int			ft_memcmp(const void *source1, const void *source2, size_t n);
 char		*ft_strnstr(const char *haystack, const char *needle, size_t len);
-int			ft_atoi(const char *number);
+long		ft_atoi(const char *number);
 void		*ft_calloc(size_t nelements, size_t byte_size);
 char		*ft_strdup(const char *source);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
